# docs/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ingest/sessions/{id}/revoke": {
      "post": {
        "operationId": "IngestAdminController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/batches": {
      "get": {
        "operationId": "IngestAdminController_batches",
        "parameters": [
          {
            "name": "sourceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/batches/{id}/integrity": {
      "get": {
        "operationId": "IngestAdminController_batchIntegrity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/items/{id}/evidence": {
      "get": {
        "operationId": "IngestAdminController_itemEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
...
    "/api/ingest/candidates/{id}": {
      "get": {
        "operationId": "IngestAdminController_candidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidate-assets/{id}/preview": {
      "get": {
        "operationId": "IngestAdminController_candidatePreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidate-assets/{id}/original": {
      "get": {
        "operationId": "IngestAdminController_candidateOriginal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidates/{id}/matches": {
      "get": {
        "operationId": "IngestAdminController_matches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidates/{id}/link-item": {
      "post": {
        "operationId": "IngestAdminController_linkItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidates/{id}/review": {
      "post": {
        "operationId": "IngestAdminController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
    "/api/ingest/candidates/{id}/confirm": {
      "post": {
        "operationId": "IngestAdminController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "外部采集接入"
        ]
      }
    },
...
    "/api/agent-ingest/batches/{id}/candidates": {
      "post": {
        "operationId": "IngestMachineController_candidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Agent导入协议"
        ]
      }
    },
    "/api/agent-ingest/batches/{id}/seal": {
      "post": {
        "operationId": "IngestMachineController_seal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Agent导入协议"
        ]
      }
    },
    "/api/agent-ingest/batches/{id}": {
      "get": {
        "operationId": "IngestMachineController_status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Agent导入协议"
        ]
      }
    },
    "/api/agent-ingest/candidates/{id}/assets": {
      "post": {
        "operationId": "IngestMachineController_asset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Agent导入协议"
        ]
      }
    },
...
  },
  "info": {
    "title": "ToMeBoutique Workbench API",
    "description": "",
    "version": "1.0.0-rc.14",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tome_session"
      }
    },
...
  }
}